mykernel2.c

+++ pa2/mykernel2.c
@@ -53,7 +53,7 @@ void refresh_slot()
       proctab[i].alive_slot += 1;
     }
     Printf("utilization of proc %d is %f, requested is %f\n", proctab[i].pid, proctab[i].utilization, proctab[i].requested);
-    proctab[i].utilization = proctab[i].ran_slot / proctab[i].alive_slot;
+    proctab[i].utilization = (double)proctab[i].ran_slot / proctab[i].alive_slot;
   }
 }
